+++ clang-tools-extra/clang-tidy/qt/QtTidyModule.cpp
@@ -10,6 +10,7 @@
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
 #include "PortToStdCompatibleApiCheck.h"
+#include "UseUnreachableReturnCheck.h"
 
 namespace clang {
 namespace tidy {
@@ -20,6 +21,8 @@ public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
     CheckFactories.registerCheck<PortToStdCompatibleApiCheck>(
         "qt-port-to-std-compatible-api");
+    CheckFactories.registerCheck<UseUnreachableReturnCheck>(
+        "qt-use-unreachable-return");
   }
 };